Keep and move squared_difference definition and simplify test
