+++ chainer/functions/__init__.py
@@ -113,6 +113,7 @@ from chainer.functions.loss.mean_squared_error import mean_squared_error  # NOQA
 from chainer.functions.loss.negative_sampling import negative_sampling  # NOQA
 from chainer.functions.loss.sigmoid_cross_entropy import sigmoid_cross_entropy  # NOQA
 from chainer.functions.loss.softmax_cross_entropy import softmax_cross_entropy  # NOQA
+from chainer.functions.loss.squared_error import squared_difference  # NOQA
 from chainer.functions.loss.squared_error import squared_error  # NOQA
 from chainer.functions.loss.triplet import triplet  # NOQA
 from chainer.functions.loss.vae import bernoulli_nll  # NOQA
@@ -228,4 +229,3 @@ from chainer.functions.util.forget import forget  # NOQA
 
 # Aliases
 from chainer.functions.math.average import average as mean  # NOQA
-squared_difference = squared_error  # NOQA
