+++ chainer/functions/loss/squared_error.py
@@ -53,7 +53,10 @@ def squared_error(x0, x1):
 
         :func:`~chainer.functions.squared_error` and
         :func:`~chainer.functions.squared_difference` are identical functions.
+        Only with different argument names.
         They are both kept for backward compatibility.
+
+    .. seealso:: :func:`~chainer.functions.squared_difference`
 
     .. admonition:: Example
 
@@ -63,11 +66,25 @@ def squared_error(x0, x1):
         >>> x2 = np.array([5, 4, 3, 2, 1, 0]).astype(np.float32)
         >>> x2
         array([5., 4., 3., 2., 1., 0.], dtype=float32)
-        >>> y = F.squared_error(x1, x2)  # Or F.squared_difference(x1, x2)
+        >>> y = F.squared_error(x1, x2)
         >>> y.shape
         (6,)
         >>> y.array
         array([25.,  9.,  1.,  1.,  9., 25.], dtype=float32)
 
+    .. seealso:: :func:`~chainer.functions.squared_difference`
+
     """
     return SquaredError().apply((x0, x1))[0]
+
+
+def squared_difference(x1, x2):
+    """Squared difference function.
+
+    This functions is identical to :func:`~chainer.functions.squared_error`
+    except for the names of the arguments.
+
+    .. seealso:: :func:`~chainer.functions.squared_error`
+
+    """
+    return squared_error(x1, x2)
